+++ src/Schema/Directives/DeleteDirective.php
@@ -92,6 +92,7 @@ GRAPHQL;
             }
         } else {
             /** @var \Illuminate\Database\Eloquent\Model $related */
+            // @phpstan-ignore-next-line Relation&Builder mixin not recognized
             $related = $relation->make();
             $related::destroy($idOrIds);
         }