+++ CHANGELOG.md
@@ -17,10 +17,15 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Handle content types `application/graphql` and `application/x-www-form-urlencoded` properly https://github.com/nuwave/lighthouse/pull/1424
 - Mark directives that can be used more than once per location as `repeatable` https://github.com/nuwave/lighthouse/pull/1529
 - Allow configuring global field middleware directives in `config/lighthouse.php` https://github.com/nuwave/lighthouse/pull/1533
-- Add laravel echo compatible subscription broadcaster. https://github.com/nuwave/lighthouse/pull/1370
+- Add custom attributes to validations https://github.com/nuwave/lighthouse/pull/1628
+- Add new directive interface `FieldBuilderDirective` https://github.com/nuwave/lighthouse/pull/1636
+- Add `@whereAuth` directive for filtering a field based on authenticated user https://github.com/nuwave/lighthouse/pull/1636
+- Use the `@trim` directive on fields to sanitize all input strings https://github.com/nuwave/lighthouse/pull/1641
+- Add Laravel Echo compatible subscription broadcaster https://github.com/nuwave/lighthouse/pull/1370
 
 ### Changed
 
+- Use `Illuminate\Support\Carbon` instead of `Carbon\Carbon`
 - `\Nuwave\Lighthouse\Exceptions\ValidationException` no longer extends `\Illuminate\Validation\ValidationException` https://github.com/nuwave/lighthouse/pull/1185
 - Move validation related classes into namespace `Nuwave\Lighthouse\Validation` https://github.com/nuwave/lighthouse/pull/1185
 - Run `ArgDirectives` in distinct phases: Sanitize, Validate, Transform https://github.com/nuwave/lighthouse/pull/1185
@@ -50,7 +55,10 @@ You can find and compare releases at the [GitHub release page](https://github.co
 - Allow subscriptions without named operations, base channels on the field name
 - Set `lighthouse.debug` config through env `LIGHTHOUSE_DEBUG` https://github.com/nuwave/lighthouse/pull/1592
 - Test helper `multipartGraphQL` now accepts arrays instead of JSON strings https://github.com/nuwave/lighthouse/pull/1615/
-- Use DateTime::ATOM for DateTimeTZ ISO 8601 compatibility https://github.com/nuwave/lighthouse/pull/1622
+- Use `DateTime::ATOM` for DateTimeTZ ISO 8601 compatibility https://github.com/nuwave/lighthouse/pull/1622
+- Split `ProvidesRules` interface into `ArgumentValidation` and `ArgumentSetValidation` https://github.com/nuwave/lighthouse/pull/1628
+- Update to PHP 8 compatible mll-lab/graphql-php-scalars 4 https://github.com/nuwave/lighthouse/pull/1639
+- Add `TrimDirective` to the default `field_middleware` config in `lighthouse.php` https://github.com/nuwave/lighthouse/pull/1641
 
 ### Removed
 
@@ -73,6 +81,7 @@ You can find and compare releases at the [GitHub release page](https://github.co
 ### Fixed
 
 - Prefix complex conditions with table name to avoid ambiguous SQL https://github.com/nuwave/lighthouse/pull/1530
+- Merge type interfaces when extending type https://github.com/nuwave/lighthouse/pull/1635
 
 ## 4.18.0